Move renderApiException to a seperate more logical file
