+++ app/Http/Middleware/FormatJsonApiError.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Middleware;
 
-use App\Exceptions\ErrorReporting;
+use App\Exceptions\JsonApiErrorRenderer;
 use Closure;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
@@ -11,7 +11,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\ServiceUnavailableHttpException;
 
 /**
- * Convert failures that never reach ErrorReporting::renderApiException()
+ * Convert failures that never reach JsonApiErrorRenderer::render()
  * (which renders all other v1 API errors) into JSON:API 503 documents:
  * database failures thrown by middleware later in this stack, and server
  * errors that Laravel's pipeline already rendered into the health
@@ -40,7 +40,7 @@ class FormatJsonApiError
 
     protected function serviceUnavailable(): Response
     {
-        return ErrorReporting::renderApiException(
+        return JsonApiErrorRenderer::render(
             new ServiceUnavailableHttpException(null, 'A required service is unavailable.'),
         );
     }
